+++ argilla-server/tests/unit/api/handlers/v1/datasets/records/records_bulk/test_create_dataset_records_bulk.py
@@ -12,14 +12,19 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from uuid import UUID
-
 import pytest
-from argilla_server.enums import DatasetStatus, QuestionType, ResponseStatus, SuggestionType
-from argilla_server.models.database import Record, Response, Suggestion, User
+
+from uuid import UUID
 from httpx import AsyncClient
+from fastapi.encoders import jsonable_encoder
 from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import AsyncSession
+
+from argilla_server.jobs.queues import HIGH_QUEUE
+from argilla_server.webhooks.v1.enums import RecordEvent
+from argilla_server.webhooks.v1.records import build_record_event
+from argilla_server.models.database import Record, Response, Suggestion, User
+from argilla_server.enums import DatasetStatus, QuestionType, ResponseStatus, SuggestionType
 
 from tests.factories import (
     DatasetFactory,
@@ -32,6 +37,8 @@ from tests.factories import (
     ImageFieldFactory,
     TextQuestionFactory,
     ChatFieldFactory,
+    CustomFieldFactory,
+    WebhookFactory,
 )
 
 
@@ -472,6 +479,32 @@ class TestCreateDatasetRecordsBulk:
                 ],
             },
         )
+
+        assert response.status_code == 422
+        assert (await db.execute(select(func.count(Record.id)))).scalar_one() == 0
+
+    async def test_create_dataset_records_bulk_with_chat_field_with_non_dicts(
+        self, db: AsyncSession, async_client: AsyncClient, owner_auth_header: dict
+    ):
+        dataset = await DatasetFactory.create(status=DatasetStatus.ready)
+
+        await ChatFieldFactory.create(name="chat", dataset=dataset)
+        await LabelSelectionQuestionFactory.create(dataset=dataset)
+
+        response = await async_client.post(
+            self.url(dataset.id),
+            headers=owner_auth_header,
+            json={
+                "items": [
+                    {
+                        "fields": {
+                            "chat": "invalid",
+                        },
+                    },
+                ],
+            },
+        )
+
         assert response.status_code == 422
         assert (await db.execute(select(func.count(Record.id)))).scalar_one() == 0
 
@@ -543,6 +576,11 @@ class TestCreateDatasetRecordsBulk:
                         },
                         {
                             "loc": ["body", "items", 0, "fields", "chat"],
+                            "msg": "value is not a valid dict",
+                            "type": "type_error.dict",
+                        },
+                        {
+                            "loc": ["body", "items", 0, "fields", "chat"],
                             "msg": "str type expected",
                             "type": "type_error.str",
                         },
@@ -550,4 +588,96 @@ class TestCreateDatasetRecordsBulk:
                 },
             }
         }
+        assert (await db.execute(select(func.count(Record.id)))).scalar_one() == 0
+
+    async def test_create_dataset_records_bulk_enqueue_webhook_record_created_events(
+        self, db: AsyncSession, async_client: AsyncClient, owner_auth_header: dict
+    ):
+        dataset = await DatasetFactory.create(status=DatasetStatus.ready)
+        await TextFieldFactory.create(name="prompt", dataset=dataset)
+        await TextQuestionFactory.create(name="text-question", dataset=dataset)
+
+        webhook = await WebhookFactory.create(events=[RecordEvent.created])
+
+    async def test_create_dataset_records_bulk_with_custom_field_values(
+        self, db: AsyncSession, async_client: AsyncClient, owner_auth_header: dict
+    ):
+        dataset = await DatasetFactory.create(status=DatasetStatus.ready)
+
+        await CustomFieldFactory.create(name="custom", dataset=dataset)
+        await LabelSelectionQuestionFactory.create(dataset=dataset)
+
+        response = await async_client.post(
+            self.url(dataset.id),
+            headers=owner_auth_header,
+            json={
+                "items": [
+                    {
+                        "fields": {
+                            "prompt": "Does exercise help reduce stress?",
+                            "custom": {"a": 1, "b": 2},
+                        },
+                    },
+                    {
+                        "fields": {
+                            "prompt": "What is the best way to reduce stress?",
+                            "custom": {"c": 1, "b": 2},
+                        },
+                    },
+                    {
+                        "fields": {
+                            "custom": {"a": 1},
+                        },
+                    },
+                ],
+            },
+        )
+
+        assert response.status_code == 201
+
+        records = (await db.execute(select(Record).order_by(Record.inserted_at.asc()))).scalars().all()
+
+        event_a = await build_record_event(db, RecordEvent.created, records[0])
+        event_b = await build_record_event(db, RecordEvent.created, records[1])
+
+        assert HIGH_QUEUE.count == 2
+
+        assert HIGH_QUEUE.jobs[0].args[0] == webhook.id
+        assert HIGH_QUEUE.jobs[0].args[1] == RecordEvent.created
+        assert HIGH_QUEUE.jobs[0].args[3] == jsonable_encoder(event_a.data)
+
+        assert HIGH_QUEUE.jobs[1].args[0] == webhook.id
+        assert HIGH_QUEUE.jobs[1].args[1] == RecordEvent.created
+        assert HIGH_QUEUE.jobs[1].args[3] == jsonable_encoder(event_b.data)
+
+        assert response.status_code == 201, response.json()
+        records = (await db.execute(select(Record))).scalars().all()
+        assert len(records) == 3
+        assert records[0].fields["custom"] == {"a": 1, "b": 2}
+        assert records[1].fields["custom"] == {"c": 1, "b": 2}
+        assert records[2].fields["custom"] == {"a": 1}
+
+    async def test_create_dataset_records_bulk_with_wrong_custom_field_value(
+        self, db: AsyncSession, async_client: AsyncClient, owner_auth_header: dict
+    ):
+        dataset = await DatasetFactory.create(status=DatasetStatus.ready)
+
+        await CustomFieldFactory.create(name="custom", dataset=dataset)
+        await LabelSelectionQuestionFactory.create(dataset=dataset)
+
+        response = await async_client.post(
+            self.url(dataset.id),
+            headers=owner_auth_header,
+            json={
+                "items": [
+                    {
+                        "fields": {
+                            "custom": "invalid",
+                        },
+                    },
+                ],
+            },
+        )
+
+        assert response.status_code == 422
         assert (await db.execute(select(func.count(Record.id)))).scalar_one() == 0