+++ argilla/src/argilla/records/_dataset_records.py
@@ -63,40 +63,39 @@ class DatasetRecordsIterator:
         self.__with_responses = with_responses
         self.__with_vectors = with_vectors
         self.__records_batch = []
-        self.__limit = 1 if limit == 0 else limit
+        self.__limit = limit
+
+        if self.__limit is not None and self.__limit <= 0:
+            warnings.warn(f"Limit {self.__limit} is invalid: must be greater than 0. Setting limit to 1.")
+            self.__limit = 1
 
         if self.__limit is not None and self.__limit < self.__batch_size:
-            self.__batch_size = limit
+            self.__batch_size = self.__limit
+
 
     def __iter__(self):
         return self
 
     def __next__(self) -> Record:
-        if self._limit_reached():
-            raise StopIteration()
-
-        if not self._has_local_records():
+        if self._no_records():
             self._fetch_next_batch()
-            if not self._has_local_records():
-                raise StopIteration()
 
         return self._next_record()
 
     def _limit_reached(self) -> bool:
         if self.__limit is None:
             return False
-        return self.__limit == 0
+        return self.__limit <= 0
 
     def _next_record(self) -> Record:
+        if self._limit_reached() or self._no_records():
+            raise StopIteration()
+
         record = self.__records_batch.pop(0)
-
-        if self.__limit is not None:
-            self.__limit -= 1
-
         return record
 
-    def _has_local_records(self) -> bool:
-        return len(self.__records_batch) > 0
+    def _no_records(self) -> bool:
+        return len(self.__records_batch) <= 0
 
     def _fetch_next_batch(self) -> None:
         self.__records_batch = list(self._list())
@@ -134,7 +133,7 @@ class DatasetRecordsIterator:
         return [record_model for record_model, _ in search_items]
 
     def _is_search_query(self) -> bool:
-        return bool(self.__query and (self.__query.query or self.__query.filter))
+        return self.__query.has_search()
 
     def to_list(self, flatten: bool) -> List[Dict[str, Any]]:
         return GenericIO.to_list(records=list(self), flatten=flatten)