+++ argilla/src/argilla/records/_resource.py
@@ -183,6 +183,7 @@ class Record(Resource):
         serialized_responses = [response.serialize() for response in self.__responses]
         serialized_model["responses"] = serialized_responses
         serialized_model["suggestions"] = serialized_suggestions
+
         return serialized_model
 
     def to_dict(self) -> Dict[str, Dict]:
@@ -202,6 +203,7 @@ class Record(Resource):
         responses = self.responses.to_dict()
         vectors = self.vectors.to_dict()
 
+        # TODO: Review model attributes when to_dict and serialize methods are unified
         return {
             "id": id,
             "fields": fields,
@@ -268,8 +270,9 @@ class Record(Resource):
         )
 
         # set private attributes
-        instance._model.id = model.id
-        instance._model.status = model.status
+        instance._dataset = dataset
+        instance._model = model
+
         # Responses and suggestions are computed separately based on the record model
         instance.responses.from_models(model.responses)
         instance.suggestions.from_models(model.suggestions)
@@ -297,7 +300,18 @@ class RecordFields(dict):
         self.record = record
 
     def to_dict(self) -> dict:
-        return {key: cast_image(value) if self._is_image(key) else value for key, value in self.items()}
+        fields = {}
+
+        for key, value in self.items():
+            if value is None:
+                continue
+            elif self._is_image(key):
+                fields[key] = cast_image(value)
+            elif self._is_chat(key):
+                fields[key] = [message.model_dump() if not isinstance(message, dict) else message for message in value]
+            else:
+                fields[key] = value
+        return fields
 
     def __getitem__(self, key: str) -> FieldValue:
         value = super().__getitem__(key)
@@ -307,6 +321,11 @@ class RecordFields(dict):
         if not self.record.dataset:
             return False
         return self.record.dataset.settings.schema[key].type == "image"
+
+    def _is_chat(self, key: str) -> bool:
+        if not self.record.dataset:
+            return False
+        return self.record.dataset.settings.schema[key].type == "chat"
 
 
 class RecordMetadata(dict):